some docker fixes

+++ start-service.sh
@@ -1,10 +1,19 @@
 #!/bin/bash
 set -x
 set -e
-docker stop texla-api
-docker rm texla-api
+if [[ $(docker ps -q --filter name=texla-api$ | wc -l) -gt 0 ]]
+then
+    docker stop texla-api
+fi
+
+if [[ $(docker ps -aq --filter name=texla-api$ | wc -l) -gt 0 ]]
+then
+    docker rm texla-api
+fi
+
 git pull origin master
 git checkout master
+
 docker build -t wikitolearn/texla-api --rm .
 docker run \
   --name texla-api \
